Split into server and client modules, with appropriate public and private files and methods.
